competitors_scraper/nri/requirements.txt
selenium>=4.1.0
beautifulsoup4>=4.10.0
openai>=1.0.0
python-dotenv>=0.19.2
tqdm>=4.62.3
pandas>=1.3.5
requests>=2.27.1
webdriver-manager>=3.5.2
lxml>=4.7.1
regex>=2022.1.18
httpx>=0.23.0
tenacity>=8.0.1
